refactor: domain, controller 패키지 분리
