+++ src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
