+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain;
 
 public record ReservationTime(
         Long id,
